Fix bug for computing setId

+++ internal/provider/resource_computing.go
@@ -195,7 +195,7 @@ func resourceComputingCreate(ctx context.Context, d *schema.ResourceData, meta i
 		return diag.Errorf("Error when calling `CoreWarehouseServiceApi.CreateWarehouse_Computing``: %s\n", r.Status)
 	}
 	// response from `CreateWarehouse`: CommonDescribeJobResponse
-	d.SetId(resp.GetId())
+	d.SetId(resp.GetResourceIds()[0])
 	if errRefresh := waitJobComplete(ctx, apiClient.CoreJobServiceApi, resp.GetId()); errRefresh != nil {
 		return diag.Errorf(errRefresh.Error())
 	}
@@ -320,19 +320,19 @@ func resourceComputingDelete(ctx context.Context, d *schema.ResourceData, meta i
 		return diag.Errorf(COMPUTING_ID + " not found! ")
 	}
 
-	//False := false
-	//resp1, r1, err1 := apiClient.CoreServiceApi.StopService(ctx, resourceId).Body(cloudmgr.CoreStopServiceRequest{
-	//	Force: &False,
-	//}).Execute()
-	//if err1 != nil {
-	//	return diag.Errorf("Error when calling computing `CoreServiceApi.StopService``: %v\n", err1)
-	//}
-	//if r1.StatusCode != 200 {
-	//	return diag.Errorf("Stop resource computing fail with %d . ", r1.StatusCode)
-	//}
-	//if errRefresh := waitJobComplete(ctx, apiClient.CoreJobServiceApi, resp1.GetId()); errRefresh != nil {
-	//	return diag.Errorf(errRefresh.Error())
-	//}
+	False := false
+	resp1, r1, err1 := apiClient.CoreServiceApi.StopService(ctx, resourceId).Body(cloudmgr.CoreStopServiceRequest{
+		Force: &False,
+	}).Execute()
+	if err1 != nil {
+		return diag.Errorf("Error when calling computing `CoreServiceApi.StopService``: %v\n", err1)
+	}
+	if r1.StatusCode != 200 {
+		return diag.Errorf("Stop resource computing fail with %d . ", r1.StatusCode)
+	}
+	if errRefresh := waitJobComplete(ctx, apiClient.CoreJobServiceApi, resp1.GetId()); errRefresh != nil {
+		return diag.Errorf(errRefresh.Error())
+	}
 
 	resp, r, err := apiClient.CoreServiceApi.DeleteService(ctx, resourceId).Execute()
 	if err != nil {
